add CKEditor on cultural notes and into popover, add delete confirm on designer unit detail page, add media view on student song page

+++ src/app/components/unit-detail.component.ts
@@ -53,11 +53,14 @@ export class UnitDetailComponent implements OnInit {
         );
     }
 
-    deleteSong(song: Song): void {
-        this.courseService.deleteSong(song)
-            .then(() => {
-                this.songs = this.songs.filter(x => x != song);
-            })
+    deleteSong(song: Song, event: any): void {
+        event.preventDefault();
+        if(confirm("Are you sure you want to delete " + song.title + "?")) {
+            this.courseService.deleteSong(song)
+                .then(() => {
+                    this.songs = this.songs.filter(x => x != song);
+                })
+        }
     }
 
     createSong(): void {
